Debugged Lua Quaternion
The way we had previously implemented the Lua Quaternion class was using the "Metatables and Metamethods" documentation on the Lua website, but this did not properly link object-oriented functions to be called with a colon (:) operator.

Refactored the Quaternion class to use an inline metatable, which lets us use the colon operator.

All Lua tests now pass successfully.