Implemented Swift meaningfulLineCount function
I had to implement this function in a way that I think is synchronous. I tried doing it the async way, but it won't compile. Everyone online says there's some kind of bug with the way URLSession is implemented on non-Macs (no one in my group has a Mac), but URLSession is the only way I know of to properly implement this asynchronously.